requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiofiles
pydantic
streamlit
opencv-python
python-dotenv
pymysql
